Cloud Function/requirement.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==2.4.0
DateTime==4.4
google-cloud-error-reporting==1.5.3
google-auth==2.8.0